Test with tracetest instead of oteltest in new test mod
