+++ instrumentation/gopkg.in/macaron.v1/otelmacaron/go.mod
@@ -12,7 +12,6 @@ require (
 	go.opentelemetry.io/contrib v0.22.0
 	go.opentelemetry.io/contrib/propagators v0.22.0
 	go.opentelemetry.io/otel v1.0.0-RC2
-	go.opentelemetry.io/otel/oteltest v1.0.0-RC2
 	go.opentelemetry.io/otel/trace v1.0.0-RC2
 	gopkg.in/macaron.v1 v1.4.0
 )
